+++ src/main/java/ru/job4j/dreamjob/service/SimpleVacancyService.java
@@ -2,6 +2,8 @@ package ru.job4j.dreamjob.service;
 
 import net.jcip.annotations.ThreadSafe;
 import org.springframework.stereotype.Service;
+import ru.job4j.dreamjob.dto.FileDto;
+import ru.job4j.dreamjob.model.File;
 import ru.job4j.dreamjob.model.Vacancy;
 import ru.job4j.dreamjob.repository.VacancyRepository;
 
@@ -13,24 +15,47 @@ import java.util.Optional;
 public class SimpleVacancyService implements VacancyService {
 
     private final VacancyRepository vacancyRepository;
+    private final FileService fileService;
 
-    private SimpleVacancyService(VacancyRepository vacancyRepository) {
+    private SimpleVacancyService(VacancyRepository vacancyRepository, FileService fileService) {
         this.vacancyRepository = vacancyRepository;
+        this.fileService = fileService;
     }
 
     @Override
-    public Vacancy save(Vacancy vacancy) {
+    public Vacancy save(Vacancy vacancy, FileDto image) {
+        saveNewFile(vacancy, image);
         return vacancyRepository.save(vacancy);
+    }
+
+    private void saveNewFile(Vacancy vacancy, FileDto image) {
+        File file = fileService.save(image);
+        vacancy.setFileId(file.getId());
     }
 
     @Override
     public boolean deleteById(int id) {
-       return vacancyRepository.deleteById(id);
+        Optional<Vacancy> fileOptional = findById(id);
+        boolean result = false;
+        if (fileOptional.isPresent()) {
+            vacancyRepository.deleteById(id);
+            fileService.deleteById(fileOptional.get().getFileId());
+            result = true;
+        }
+       return result;
     }
 
     @Override
-    public boolean update(Vacancy vacancy) {
-        return vacancyRepository.update(vacancy);
+    public boolean update(Vacancy vacancy, FileDto image) {
+        boolean isNewFileEmpty = image.getContent().length == 0;
+        if (isNewFileEmpty) {
+            return vacancyRepository.update(vacancy);
+        }
+        int oldFileId = vacancy.getFileId();
+        saveNewFile(vacancy, image);
+        boolean isUpdated = vacancyRepository.update(vacancy);
+        fileService.deleteById(oldFileId);
+        return isUpdated;
     }
 
     @Override
